read length to int

+++ src/HexEditor.Core/Model/IBinaryBuffer.cs
@@ -4,7 +4,7 @@ public interface IBinaryBuffer : IAsyncDisposable
 {
 	bool TryRead(Span<byte> buffer, long offset, int length);
 
-	ValueTask CopyToAsync(Memory<byte> destination, long offset, long length, CancellationToken cancellationToken);
+	ValueTask CopyToAsync(Memory<byte> destination, long offset, int length, CancellationToken cancellationToken);
 
 	long Length { get; }
